This implements the representation mentioned in #10092, storing keys and values in dense ordered arrays, with a sparse(r) index vector. There are a couple TODOs in the code. This is expected to be a bit slower for deletion-heavy workloads, but I haven't benchmarked that yet.

My benchmark code is at https://gist.github.com/JeffBezanson/ea224aa45e2242ca3ee9 .
I took some benchmarks from past Dict performance issues, and added a couple more. The performance characteristics are interesting. Highlights:
- Simply iterating over a Dict is 10x faster, as I decided to make `start` squeeze out deleted items first.
- `rehash!` doesn't need to reallocate the keys and values arrays at all in theory, but we might need to in order to fix the problem of finalizers that remove keys during rehashing.
- Seems to be faster for String keys, but unfortunately much slower for Int keys.

Profiling indicates that a huge amount of insertion time is spent on the line

```
        elseif si > 0 && isequal(key, keys[si])
```

My best guess is that this is because `keys[si]` is random access, where before we iterated through the slots and keys arrays in order. But if pointer indirection is involved anyway (e.g. String keys) we get a significant net speedup.