@StefanKarpinski 2015-02-07
Sorry, didn't see the PR before writing that comment. This is pretty fascinating. If it's a win whenever there's pointer indirection, it's clearly a good tradeoff for Python, where that's almost always the case.