@JeffBezanson 2015-02-07
Really big gains are possible if you need the ordering feature. For example with this change, `setdiff!` can be rewritten as follows:

```
function setdiff(a, b)
    args_type = promote_type(eltype(a), eltype(b))
    bset = Set(b)
    seen = Set{args_type}()
    for a_elem in a
        !in(a_elem, bset) && push!(seen, a_elem)
    end
    collect(seen)
end
```

And I get:

```
setdiff:
elapsed time: 0.074992577 seconds (42 MB allocated, 2.35% gc time in 2 pauses with 0 full sweep)
```

which is 6x faster than master, and almost 10x faster than 0.3.