@StefanKarpinski 2015-02-07
This seems like a fairly big win overall – if we can figure out a clever way to make the Int case (immediate keys in general) faster, then I would say that switching to this would be well worth it. Another benefit is that it's one less data structure to have – no need for a separate ordered dict implementation.