@StefanKarpinski 2015-02-07
So, am I correct that the thing that's slow here for dicts of Ints is calling `copy` on them? That makes it seem like we really should be copying `Dict` objects more efficiently – why not just copy their internal arrays directly? Is the idea behind the current implementation of dict copying that rebuilding them is an equivalent amount of work to copying the internals and then rehashing?

That aside, I guess the real point of the benchmark is that inserting into an Int dict is 30% slower and iterating over an Int dict is also slower, making copy, which does both, a total of 70% slower.