@JeffBezanson 2015-02-07
Iterating over any dict, including Int dicts, is 10x faster since it becomes equivalent to iterating over two arrays. Mostly assigning is slower.

Yes, we could definitely copy more efficiently.