@timholy 2015-02-08
I rarely iterate over Dicts, so I'm curious to know how other workloads fare. For example, a way I often use Dicts is as "flexible sparse matrices," i.e.,
```
cachekey(i::Int, j::Int) = i <= j ? Pair(i, j) : Pair(j, i)
cache[key] = dot(vectors[i], vectors[j])
```
and sometime later if `vectors[j]` gets eaten, `delete!(cache, key)` for all keys associated with `j`.

I guess I should just check this branch out and benchmark it, but I won't have time for that for the next several days.