@JeffBezanson 2015-02-08
If you send me a test case I will happily benchmark it!