@tkelman 2015-02-08
Some of the CI jobs appear to be running out of memory here