@quinnj 2015-02-08
```julia
julia> r = Dict(1=>"hey",2=>"ho")
Dict{Int64,ASCIIString} with 2 entries:
  1 => "hey"
  2 => "ho"

julia> r[1]
"hey"

julia> r[2]
"ho"

julia> delete!(r,1)
Dict{Int64,ASCIIString} with 1 entry:Error showing value of type Dict{Int64,ASCIIString}:
ERROR: UndefRefError: access to undefined reference
 in showdict at dict.jl:81
 in writemime at replutil.jl:34
 in display at REPL.jl:105
 in display at REPL.jl:108
 in display at multimedia.jl:149
 in print_response at REPL.jl:127
 in print_response at REPL.jl:112
 in anonymous at REPL.jl:588

julia> r
Dict{Int64,ASCIIString} with 1 entry:Error showing value of type Dict{Int64,ASCIIString}:
ERROR: UndefRefError: access to undefined reference
 in showdict at dict.jl:81
 in writemime at replutil.jl:34
 in display at REPL.jl:105
 in display at REPL.jl:108
 in display at multimedia.jl:149
 in print_response at REPL.jl:127
 in print_response at REPL.jl:112
 in anonymous at REPL.jl:588
```