@bfredl 2015-02-08
Intresting work! However, in a use case, where dict elements are deleted when going "down" in a stack and then later added back in reverse order, when going back up the stack (a simple back-tracking algoritm), this causes a (worst-case) 3x slow-down, but I suppose this is an unavoidable consequence from the strict ordering semantics. (if an element are removed and then immediately readded, it must be pushed on front right?)
code+data [here](https://gist.github.com/bfredl/66776b24965b08d33d26) (sorry for messy code, it was an assignment where only speed mattered, not code readability. Once it was faster than the c++/STL implementation I stopped working on it :) )