@JeffBezanson 2015-02-08
Thanks for the test case, that's very helpful!