@JeffBezanson 2015-02-08
Ok, I was able to get back most of the performance by removing the `sizehint`, which was causing excess memory allocation and initialization overhead. Seems not to be worth it.