@bfredl 2015-02-09
Nice! Still somewhat slower, but I guess this is an extreme case of (structural) mutation relative to access....