@timholy 2015-02-11
Sorry for being slow, but perhaps this will be useful:
```julia
function initialize(N, nnbrs)
    nbrs_list = [Set{Int}() for i = 1:N]
    dp = Dict{Pair{Int,Int},Float64}()
    for i = 1:N
        nnbr = rand(nnbrs)
        nbrs = nbrs_list[i]
        for j = 1:nnbr
            nbr = rand(1:N)
            push!(nbrs, nbr)
            dp[cachekey(i,j)] = rand()
        end
    end
    nbrs_list, dp
end

function eliminate!(nbrs_list, dp)
    N = length(nbrs_list)
    order = randperm(N)
    for i in order
        nbrs = nbrs_list[i]
        for j in nbrs
            delete!(dp, cachekey(i,j))
        end
        empty!(nbrs)
    end
end

cachekey(i, j) = i <= j ? Pair(i, j) : Pair(j, i)

nnbrs = 1:20
nbrs_list, dp = initialize(5, nnbrs)
eliminate!(nbrs_list, dp)
@time 1

N = 10^5
@time nbrs_list, dp = initialize(N, nnbrs)
@time eliminate!(nbrs_list, dp)
```