@hayd 2015-05-22
Is it possible this could make 0.4? As this is awesome.

Does the int-case have to be special-cased for performance (e.g. using the old code)? Note: intset is special-cased (see #10065).