@kmsquire 2015-05-25
If this doesn't make 0.4, it might be worth replacing OrderedDict in
DataStructures.jl with this version.

On Friday, May 22, 2015, Andy Hayden <notifications@github.com> wrote:

> Is it possible this could make 0.4? As this is awesome.
>
> Does the int-case have to be special-cased for performance (e.g. using the
> old code)? Note: intset is special-cased (see #10065
> <https://github.com/JuliaLang/julia/pull/10065>).
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/JuliaLang/julia/pull/10116#issuecomment-104772110>.
>