@JeffBezanson 2015-05-26
I'm starting to think we should go with this. Int keys are arguably not the most important case, and they're easy to special-case if necessary. Being faster for strings, faster iteration, and not needing a separate OrderedDict seem to be worth it.