@timholy 2015-05-26
If you try that benchmark I posted, I'd be curious to know how it fares.