@JeffBezanson 2015-05-30
The timings are a bit variable, but I got this on master:

```
   2.582 microseconds (155 allocations: 10845 bytes)
 522.997 milliseconds (1099 k allocations: 184 MB, 19.66% gc time)
 182.668 milliseconds (6 allocations: 781 KB)
```

and on this branch:

```
   2.451 microseconds (155 allocations: 10845 bytes)
 502.958 milliseconds (1170 k allocations: 152 MB, 22.28% gc time)
 188.579 milliseconds (6 allocations: 781 KB)
```