@JeffBezanson 2015-05-30
Best time I've seen on master:

```
   2.916 microseconds (155 allocations: 10845 bytes)
 457.631 milliseconds (1099 k allocations: 184 MB, 22.73% gc time)
 172.746 milliseconds (6 allocations: 781 KB)
```

best time here:

```
   2.634 microseconds (155 allocations: 10845 bytes)
 466.899 milliseconds (1170 k allocations: 152 MB, 23.64% gc time)
 182.558 milliseconds (6 allocations: 781 KB)
```

I think it's at least safe to conclude there's no major regression.