@ScottPJones 2015-05-30
LGTM too... (and I do plan on using ordered Dicts heavily, so getting this in will be very nice)