@StefanKarpinski 2015-05-30
I'm all for making this change. Ordered Dicts are very useful. One issues this raises is that dicts are now distinguished not just by their contents but by their ordering. Should two Dicts be considered equal if they have different orders? While making Dicts ordered is non-breaking, changing their equality is not.