@JeffBezanson 2015-05-30
Very good point. I'd think that since ordering is part of the interface, equality would have to take it into account.