@mauro3 2015-05-31
I adapted the test-script I used to profile dict work I did some time ago (PR  #7348):
https://gist.github.com/mauro3/2352389f7f9618e6e1b9

It times the test suite (first number) and insertion, deletion and insertion-deletion (second number):
Master:
```
327.6ms
546.2ms
```
this PR:
```
255.9ms
493.2ms
```
So about 10-25% faster!

Also, I agree with @hayd, that equality should not depend on order, and to add a `OrderedDict` which just differs by the `==`.